마커 애니메이션 출력에 약간의 오차가 있던 문제점 개선

+++ Assets/Scripts/MarkerManager.cs
@@ -124,7 +124,7 @@ public class MarkerManager : MonoBehaviour
 
         var markerAnimation = markerObj.GetComponent<MarkerAnimator>();
         markerAnimation.SpriteList = CurrentMarkerSprites[0];
-        markerAnimation.StartTime = (float)(note.StartTime + GameManager.Instance.StartTime);
+        markerAnimation.StartTime = note.StartTime + GameManager.Instance.StartTime;
     }
 
     public void PlayClap()
